Convey

+++ Backend/controllers/user.controller.js
@@ -26,32 +26,31 @@ export const RegisterUser = async (req, res) => {
             });
         }
  
-       console.log("1. Register API Hit");
+        const userExists = await UserModel.findOne({ email });
 
-const userExists = await UserModel.findOne({ email });
-console.log("2. User checked");
-
-const otp = Math.floor(1000 + Math.random() * 9000).toString();
-console.log("3. OTP Generated");
-
-const hashedPassword = await bcrypt.hash(password, 11);
-console.log("4. Password Hashed");
-
-await OtpModel.deleteOne({ email });
-console.log("5. Old OTP Deleted");
-
-await OtpModel.create({
-    fullname,
-    email,
-    password: hashedPassword,
-    profile,
-    otp,
-    otpExpiry: new Date(Date.now() + 5 * 60 * 1000),
-});
-console.log("6. OTP Saved");
-
-await sendOtpMail(email, fullname, otp);
-console.log("7. Email Sent");
+        if (userExists) {
+            return res.status(409).json({
+                message: "User already exists",
+                success: false,
+            });
+        }
+ 
+        const otp = Math.floor(1000 + Math.random() * 9000).toString();
+ 
+        const hashedPassword = await bcrypt.hash(password, 11);
+ 
+        await OtpModel.deleteOne({ email });
+ 
+        await OtpModel.create({
+            fullname,
+            email,
+            password: hashedPassword,
+            profile,
+            otp,
+            otpExpiry: new Date(Date.now() + 5 * 60 * 1000),  
+        });
+ 
+        await sendOtpMail(email, fullname, otp);
 
         return res.status(200).json({
             success: true,
